requirements.txt
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn[standard]==0.29.0
jinja2==3.1.3
sqlalchemy==2.0.30
mysql-connector-python==8.4.0
requests==2.31.0
python-multipart==0.0.9
redis==5.0.4